Added tests, Schools, Admin control, and validation

+++ private/controllers/Schools.php
@@ -21,10 +21,14 @@ class Schools extends Controller
 		$crumbs[] = ['Dashboard',''];
 		$crumbs[] = ['Schools','schools'];
 
-		$this->view('schools',[
-			'crumbs'=>$crumbs,
-			'rows'=>$data
-		]);
+		if(Auth::access('super_admin')){
+			$this->view('schools',[
+				'crumbs'=>$crumbs,
+				'rows'=>$data
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 
 	public function add()
@@ -36,7 +40,7 @@ class Schools extends Controller
 		}
 
 		$errors = array();
-		if(count($_POST) > 0)
+		if(count($_POST) > 0 && Auth::access('super_admin'))
  		{
 
 			$school = new School();
@@ -58,11 +62,15 @@ class Schools extends Controller
 		$crumbs[] = ['Schools','schools'];
 		$crumbs[] = ['Add','schools/add'];
 
-		$this->view('schools.add',[
-			'errors'=>$errors,
-			'crumbs'=>$crumbs,
-			
-		]);
+		if(Auth::access('super_admin')){
+			$this->view('schools.add',[
+				'errors'=>$errors,
+				'crumbs'=>$crumbs,
+				
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 
 	public function edit($id = null)
@@ -76,7 +84,7 @@ class Schools extends Controller
 		$school = new School();
 
 		$errors = array();
-		if(count($_POST) > 0)
+		if(count($_POST) > 0 && Auth::access('super_admin'))
  		{
 
 			if($school->validate($_POST))
@@ -97,11 +105,16 @@ class Schools extends Controller
 		$crumbs[] = ['Schools','schools'];
 		$crumbs[] = ['Edit','schools/edit'];
 
-		$this->view('schools.edit',[
-			'row'=>$row,
-			'errors'=>$errors,
-			'crumbs'=>$crumbs,
-		]);
+		if(Auth::access('super_admin')){
+
+			$this->view('schools.edit',[
+				'row'=>$row,
+				'errors'=>$errors,
+				'crumbs'=>$crumbs,
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 
 	public function delete($id = null)
@@ -116,7 +129,7 @@ class Schools extends Controller
 
 		$errors = array();
 
-		if(count($_POST) > 0)
+		if(count($_POST) > 0 && Auth::access('super_admin'))
  		{
  
  			$school->delete($id);
@@ -130,10 +143,14 @@ class Schools extends Controller
 		$crumbs[] = ['Schools','schools'];
 		$crumbs[] = ['Delete','schools/delete'];
 
-		$this->view('schools.delete',[
-			'row'=>$row,
- 			'crumbs'=>$crumbs,
-		]);
+		if(Auth::access('super_admin')){
+			$this->view('schools.delete',[
+				'row'=>$row,
+	 			'crumbs'=>$crumbs,
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 	
 }
